// apps/web/src/components/ui/buttons/button/types.ts
export enum ButtonSize {
  sm = 'sm',
  base = 'base',
  lg = 'lg',
  iconSm = 'iconSm',
  iconBase = 'iconBase',
  iconLg = 'iconLg',
}

export enum ButtonVariant {
  default = 'default',
  outline = 'outline',
  ghost = 'ghost',
  destructive = 'destructive',
  secondary = 'secondary',
  mono = 'mono',
}